+++ flowmachine/flowmachine/features/network/total_network_objects.py
@@ -153,24 +153,22 @@ class TotalNetworkObjects(GeoDataMixin, Query):
         ]
 
     def _make_query(self):
-        cols = ",".join(get_columns_for_level(self.network_object))
-        group_cols = ",".join(
-            get_columns_for_level(self.level, self.joined.column_name)
-        )
-        sql = """
-        SELECT {group_cols}, COUNT(*) as value,
+        cols = get_columns_for_level(self.network_object)
+        group_cols = get_columns_for_level(self.level, self.joined.column_name)
+        for column in group_cols:
+            if column in cols:
+                cols.remove(column)
+        cols_str = ",".join(cols)
+        group_cols_str = ",".join(group_cols)
+        sql = f"""
+        SELECT {group_cols_str}, COUNT(*) as value,
              datetime FROM
-              (SELECT DISTINCT {group_cols}, {cols}, datetime FROM           
-                (SELECT {group_cols}, {cols}, date_trunc('{total_by}', x.datetime) AS datetime
-                FROM ({etu}) x) y) _
-            GROUP BY {group_cols}, datetime
-            ORDER BY {group_cols}, datetime
-        """.format(
-            total_by=self.total_by,
-            etu=self.joined.get_query(),
-            cols=cols,
-            group_cols=group_cols,
-        )
+              (SELECT DISTINCT {group_cols_str}, {cols_str}, datetime FROM           
+                (SELECT {group_cols_str}, {cols_str}, date_trunc('{self.total_by}', x.datetime) AS datetime
+                FROM ({self.joined.get_query()}) x) y) _
+            GROUP BY {group_cols_str}, datetime
+            ORDER BY {group_cols_str}, datetime
+        """
 
         return sql